Fix topic_propagation

+++ run_package.py
@@ -1,10 +1,8 @@
 import os
 from pathlib import Path
-
-from polarity_quntification.polarization_algorithms.topic_propagation import topic_propagation
-
 os.environ["METIS_DLL"] = "polarity_quntification/dll/metis.dll"
 
+from polarity_quntification.polarization_algorithms.topic_propagation import topic_propagation, read_tweets_from_file
 import polarity_quntification.polarization_algorithms.polarization_algorithms as pol
 import networkx as nx
 import polarity_quntification.partition_algorithms.partition_algorithms as pa
@@ -93,10 +91,39 @@ def compute_polarization(dG):
 network_path = Path('data/networks/retweet_networks/')
 suffix = '_threshold_largest_CC.txt'
 prefix = 'retweet_graph_'
-output_path = Path('structural_polarity_quantification_scores_kiran_dataset.csv')
-cols = ['graph_name', 'rwc_metis', 'arwc_metis', 'ebc_metis', 'gmck_metis', 'mblb_metis', 'mod_metis', 'ei_metis',
-        'extei_metis', 'cond_metis', 'rwc_rsc', 'arwc_rsc', 'ebc_rsc', 'gmck_rsc', 'mblb_rsc', 'mod_rsc', 'ei_rsc',
-        'extei_rsc', 'cond_rsc', 'size', 'ave_deg']
+#output_path = Path('structural_polarity_quantification_scores_kiran_dataset.csv')
+#cols = ['graph_name', 'rwc_metis', 'arwc_metis', 'ebc_metis', 'gmck_metis', 'mblb_metis', 'mod_metis', 'ei_metis',
+#        'extei_metis', 'cond_metis', 'rwc_rsc', 'arwc_rsc', 'ebc_rsc', 'gmck_rsc', 'mblb_rsc', 'mod_rsc', 'ei_rsc',
+#        'extei_rsc', 'cond_rsc', 'size', 'ave_deg']
+#
+#if output_path.exists():
+#    processed_graphs = pd.read_csv(output_path)['graph_name'].tolist()
+#else:
+#    processed_graphs = []
+#
+#graphs_n = len(list(network_path.iterdir())) - len(processed_graphs)
+#for edgelist_file in tqdm(network_path.iterdir(), desc='process graphs', total=graphs_n):
+#    if edgelist_file.name.endswith(suffix):
+#        graph_name = edgelist_file.name.replace(suffix, '').replace(prefix, '')
+#        if graph_name not in processed_graphs:
+#            print(graph_name)
+#            G = nx.read_weighted_edgelist(edgelist_file, delimiter=',')
+#            G.graph['edge_weight_attr'] = 'weight'
+#            edge_weight = nx.get_edge_attributes(G, 'weight')
+#            nx.set_edge_attributes(G, {k: int(v) for k, v in edge_weight.items()}, 'weight')
+#
+#            row = [graph_name] + compute_polarization(G)
+#            if output_path.exists():
+#                pd.DataFrame([row], columns=cols).to_csv(output_path, index=False, header=False, mode='a')
+#            else:
+#                pd.DataFrame([row], columns=cols).to_csv(output_path, index=False)
+
+
+tweet_path = Path('data/full_tweets/')
+output_path = Path('topic_propagation_scores_kiran_dataset.csv')
+
+base_score_names = ['directed_NMI', 'directed_AMI', 'directed_ARI', 'undirected_NMI', 'undirected_AMI', 'undirected_ARI']
+cols = ['graph_name'] + ['louvain' + s for s in base_score_names] + ['metis' + s for s in base_score_names] + ['rsc' + s for s in base_score_names]
 
 if output_path.exists():
     processed_graphs = pd.read_csv(output_path)['graph_name'].tolist()
@@ -107,18 +134,20 @@ graphs_n = len(list(network_path.iterdir())) - len(processed_graphs)
 for edgelist_file in tqdm(network_path.iterdir(), desc='process graphs', total=graphs_n):
     if edgelist_file.name.endswith(suffix):
         graph_name = edgelist_file.name.replace(suffix, '').replace(prefix, '')
-        if graph_name not in processed_graphs:
+        if graph_name not in processed_graphs and (tweet_path / f'{graph_name}.txt').exists():
             print(graph_name)
             G = nx.read_weighted_edgelist(edgelist_file, delimiter=',')
             G.graph['edge_weight_attr'] = 'weight'
             edge_weight = nx.get_edge_attributes(G, 'weight')
             nx.set_edge_attributes(G, {k: int(v) for k, v in edge_weight.items()}, 'weight')
-
-            row = [graph_name] + compute_polarization(G)
+            
+            tweets = read_tweets_from_file(tweet_path / f'{graph_name}.txt', filter_retweets=False) 
+            
+            louvain_scores , metis_scores , rsc_scores = topic_propagation(graph_name, G, tweets, network_type='retweet')            
+            row = [graph_name] + louvain_scores + metis_scores + rsc_scores
             if output_path.exists():
                 pd.DataFrame([row], columns=cols).to_csv(output_path, index=False, header=False, mode='a')
             else:
                 pd.DataFrame([row], columns=cols).to_csv(output_path, index=False)
 
 
-topic_propagation()
